+++ ledger/storage/iterate.go
@@ -56,7 +56,7 @@ func pulseNumRecordPrefix(n record.PulseNum) []byte {
 //
 // Error returned by the ProcessSlotRecords is based on iteration function
 // result or BadgerDB iterator error if any.
-func (s *Store) ProcessSlotHashes(n record.PulseNum, ifn func(it HashIterator) error) error {
+func (s *DB) ProcessSlotHashes(n record.PulseNum, ifn func(it HashIterator) error) error {
 	prefix := pulseNumRecordPrefix(n)
 
 	iopts := badger.DefaultIteratorOptions
@@ -72,7 +72,7 @@ func (s *Store) ProcessSlotHashes(n record.PulseNum, ifn func(it HashIterator) e
 }
 
 // GetSlotHashes returns array of all record's hashes in provided PulseNum.
-func (s *Store) GetSlotHashes(n record.PulseNum) ([][]byte, error) {
+func (s *DB) GetSlotHashes(n record.PulseNum) ([][]byte, error) {
 	var hashes [][]byte
 	err := s.ProcessSlotHashes(n, func(it HashIterator) error {
 		for it.Next() {
